Remove kmd sign+submit example ~duplicate
